Add functionality for tracking completed exercises
In this commit, the code was updated to provide functionality to track exercises completed by a user. A new migration was added to create a 'completed_exercises' table in the database that records which exercises a user has completed. The User model was updated with a completed_exercises method that returns the exercises completed by the user. The ExerciseController now includes a method for marking an exercise as completed when it is viewed.
A new view was also added to display a message to the user when all exercises have been completed. This will help keep users informed about their progress and motivate them to continue using the application. The necessary route and controller method were added to support this new feature.

Finally, the role field was moved back to the users table from the deleted add_role_to_users_table migration. Now, it's set to default to 'user' when a new user registers, as specified in the RegisterController.